refactor: split RTCStatsServer into multiple modules
